split the toggle setting into its own file
